+++ src/ReadestStats/Views/ImportPreviewWindow.xaml.cs
@@ -1,0 +1,9 @@
+using System.Windows;
+
+namespace ReadestStats.Views;
+
+public partial class ImportPreviewWindow : Window
+{
+    public ImportPreviewWindow() => InitializeComponent();
+    private void Confirm_Click(object sender, RoutedEventArgs e) => DialogResult = true;
+}